Add more tests for QMLSaleCartModel
- Add last request to AbstractVisualListModel
- Create DebtorPusher
- Add handling for EmptyCartError in QMLSaleCartModel
- Clean up model list in QMLSaleHomeModel
- Correct function names in SaleSqlManager
- Add more tests for QMLSaleCartModel
- Correct test names
- Drop database and close connection in DatabaseClient's destructor

+++ tests/utils/databaseclient.cpp
@@ -24,8 +24,10 @@ DatabaseClient::DatabaseClient()
 DatabaseClient::~DatabaseClient()
 {
     try {
-        DatabaseCreator(m_connection).dropDatabase(DATABASENAME);
-        disconnectFromTestDatabase();
+        if (m_connection.isOpen()) {
+            DatabaseCreator(m_connection).dropDatabase(DATABASENAME);
+            disconnectFromTestDatabase();
+        }
     } catch (DatabaseException &e) {
         qDebug() << "Exception caught:" << e.message() << e.userMessage();
     }
